requirements.txt
flask>=2.2.0
transformers>=4.36.0
soundfile>=0.12.1
threadpoolctl>=3.1.0
numpy>=1.21.0
scipy>=1.7.0
accelerate>=0.26.0
gunicorn>=21.2.0
xformers
torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0
sagemaker-inference==1.10.1
realtimetts[all]